Solution to Input Two Floats and Print Their Product

+++ Input Two Floats and Print Their Product@@6788d5012db25efec946f93e/code.c
@@ -4,7 +4,7 @@
 int main() {
     float a,b;
     scanf("%f %f",&a,&b);
-    int p=a*b;
+    float p=a*b;
     printf("Product: %.2f",p);
     return 0;
 }
